model-instance conversion smoking test passed

+++ src/main/java/tech/metavm/object/instance/rest/SelectRequestDTO.java
@@ -3,7 +3,7 @@ package tech.metavm.object.instance.rest;
 import java.util.List;
 
 public record SelectRequestDTO (
-        int typeId,
+        long typeId,
         List<String> selects,
         String condition,
         int page,
